Refactoring EventArgs
